chg: :truck: rename ops.py into decorators.py
